+++ packages/nibijs/src/chain/chain.ts
@@ -83,14 +83,21 @@ export const Chaosnet: Chain = {
 }
 
 /**
- * Sends 10 NIBI and 100_000 NUSD to the given address from the testnet faucet.
+ * Sends 10 NIBI and 100 NUSD to the given address from the testnet faucet.
  */
-export async function useFaucet(
-  address: string,
-  faucetUrl?: string,
-): Promise<Response> {
-  const amtNibi = 10
-  const amtNusd = 100_000
+export async function useFaucet({
+  address,
+  faucetUrl,
+  amtNibi,
+  amtNusd,
+}: {
+  address: string
+  faucetUrl?: string
+  amtNibi?: number
+  amtNusd?: number
+}): Promise<Response> {
+  amtNibi = amtNibi ?? 10
+  amtNusd = amtNusd ?? 100
   const coins: string[] = [
     `${(amtNibi * 1_000_000).toString()}unibi`,
     `${(amtNusd * 1_000_000).toString()}unusd`,