+++ packages/nibijs/src/test/chain.test.ts
@@ -10,10 +10,11 @@ import {
   queryChainIdWithRest,
   Testnet,
   assert,
-} from "./chain"
+} from "../chain"
+import { TEST_CHAIN } from "./helpers"
 
 describe("chain connections", () => {
-  const chain: Chain = Testnet
+  const chain: Chain = TEST_CHAIN
 
   test("testnet rpc", async () => {
     const sgClient = await SigningStargateClient.connect(chain.endptTm)